Refactor Code
