+++ forwords/navigation/MainTabNavigator.js
@@ -1,14 +1,13 @@
 import React from 'react';
 import { createStackNavigator } from 'react-navigation';
 import HomeScreen from '../screens/HomeScreen';
-import LoginScreen from '../screens/LoginScreen';
-import SettingsScreen from '../screens/SettingsScreen';
+import LoginScreen from '../screens/Authentication/LoginScreen';
 import GamePlayScreen from '../screens/GamePlay/GamePlayScreen';
-import SinglePlayerModeSelectionScreen from '../screens/SinglePlayerModeSelectionScreen';
-import RegisterScreen from '../screens/RegisterScreen';
-import InstructionScreen from '../screens/InstructionScreen';
+import SinglePlayerModeSelectionScreen from '../screens/Selection/SinglePlayerModeSelectionScreen';
+import RegisterScreen from '../screens/Authentication/RegisterScreen';
+import InstructionScreen from '../screens/GamePlay/Instructions/InstructionScreen';
 import PlaceholderScreen from '../screens/PlaceholderScreen';
-import ForgotPasswordScreen from '../screens/ForgotPasswordScreen';
+import ForgotPasswordScreen from '../screens/Authentication/ForgotPasswordScreen';
 
 const HomeStack = createStackNavigator({
   Home: HomeScreen,
@@ -30,10 +29,6 @@ const RegisterStack = createStackNavigator({
   Register: RegisterScreen,
 });
 
-const SettingsStack = createStackNavigator({
-  Settings: SettingsScreen,
-});
-
 const SinglePlayerModeSelectionStack = createStackNavigator({
   SinglePlayerModeSelection: SinglePlayerModeSelectionScreen,
 });
@@ -47,12 +42,12 @@ const InstructionStack = createStackNavigator({
 });
 
 export default createStackNavigator({
+  SinglePlayerModeSelectionStack,
   LoginStack,
   RegisterStack,
   HomeStack,
   ForgotPasswordStack,
   PlaceholderStack,
-  SettingsStack,
   SinglePlayerModeSelectionStack,
   GamePlayScreenStack,
   InstructionStack,
